# openapi.json
{
  "components": {
    "schemas": {
...
      "Boolean": {
        "type": "boolean"
      },
...
      "GetUploadFindingsResult": {
        "properties": {
          "challenge": {
            "$ref": "#/components/schemas/IReadOnlyList\u00601"
          },
          "findings": {
            "$ref": "#/components/schemas/IReadOnlyList\u00601"
          },
          "narrative": {
            "$ref": "#/components/schemas/IReadOnlyList\u00601"
          },
          "review": {
            "$ref": "#/components/schemas/IReadOnlyList\u00601"
          },
          "uploadId": {
            "$ref": "#/components/schemas/Guid"
          }
        },
        "required": [
          "uploadId",
          "findings",
          "narrative",
          "challenge",
          "review"
        ],
        "type": "object"
      },
...
      "Guid": {
        "format": "uuid",
        "type": "string"
      },
      "IFormFile": {
        "format": "binary",
        "type": "string"
      },
...
      "String": {
        "type": "string"
      },
...
      "SummarizeProgressResult": {
        "properties": {
          "hasPrevious": {
            "$ref": "#/components/schemas/Boolean"
          },
          "latestRunAt": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Nullable\u00601"
              }
            ]
          },
          "movedBackward": {
            "$ref": "#/components/schemas/IReadOnlyList\u00601"
          },
          "movedForward": {
            "$ref": "#/components/schemas/IReadOnlyList\u00601"
          },
          "pace": {
            "$ref": "#/components/schemas/String"
          },
          "paceIsPlaceholder": {
            "$ref": "#/components/schemas/Boolean"
          },
          "previousRunAt": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Nullable\u00601"
              }
            ]
          },
          "projectKey": {
            "$ref": "#/components/schemas/String"
          },
          "scoreAfter": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Nullable\u00601"
              }
            ]
          },
          "scoreBefore": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Nullable\u00601"
              }
            ]
          },
          "scoreDelta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Nullable\u00601"
              }
            ]
          }
        },
        "required": [
          "projectKey",
          "hasPrevious",
          "previousRunAt",
          "latestRunAt",
          "scoreBefore",
          "scoreAfter",
          "scoreDelta",
          "pace",
          "paceIsPlaceholder",
          "movedForward",
          "movedBackward"
        ],
        "type": "object"
      },
      "UploadFixtureResult": {
        "properties": {
          "fileName": {
            "$ref": "#/components/schemas/String"
          },
          "uploadId": {
            "$ref": "#/components/schemas/Guid"
          }
        },
        "required": [
          "uploadId",
          "fileName"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AiPMOInsight.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/ingest/upload": {
      "post": {
        "operationId": "UploadFixture",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadFixtureResult"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "tags": [
          "Ingest"
        ]
      }
    },
...
    "/api/projects/{projectKey}/progress": {
      "get": {
        "operationId": "GetProjectProgress",
        "parameters": [
          {
            "in": "path",
            "name": "projectKey",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/String"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummarizeProgressResult"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "tags": [
          "Progress"
        ]
      }
    },
...
    "/api/uploads/{uploadId}/findings": {
      "get": {
        "operationId": "GetUploadFindings",
        "parameters": [
          {
            "in": "path",
            "name": "uploadId",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Guid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUploadFindingsResult"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "tags": [
          "History"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
...
}